+++ cpp/monitor/sm_vcd_report.cpp
@@ -75,8 +75,8 @@ static uint8_t get_data_field(uint32_t data)
 static void vcd_output_if_changed_phase(ofstream& fp, uint32_t data, int pin, char symbol)
 {
     const BUS::phase_t new_value = GPIOBUS::GetPhaseRaw(data);
-    if (prev_value[pin] != static_cast<int>(new_value)) {
-        prev_value[pin] = static_cast<int>(new_value);
+    if (prev_value[pin] != static_cast<uint8_t>(new_value)) {
+        prev_value[pin] = static_cast<uint8_t>(new_value);
         fp << "s" << GPIOBUS::GetPhaseStrRaw(new_value) << " " << symbol << endl;
     }
 }
@@ -164,7 +164,7 @@ void scsimon_generate_value_change_dump(const char *filename, const data_capture
 
     uint32_t i = 0;
     while (i < capture_count) {
-        vcd_ofstream << "#" << (uint64_t)(data_capture_array[i].timestamp * ns_per_loop) << endl;
+        vcd_ofstream << "#" << (uint64_t)((double)data_capture_array[i].timestamp * ns_per_loop) << endl;
         vcd_output_if_changed_bool(vcd_ofstream, data_capture_array[i].data, PIN_BSY, SYMBOL_PIN_BSY);
         vcd_output_if_changed_bool(vcd_ofstream, data_capture_array[i].data, PIN_SEL, SYMBOL_PIN_SEL);
         vcd_output_if_changed_bool(vcd_ofstream, data_capture_array[i].data, PIN_CD, SYMBOL_PIN_CD);
